商品が購入されていないか確認する記述を追加

+++ app/controllers/items_controller.rb
@@ -41,7 +41,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if current_user.id == @item.user_id
+    if @item.bought_item_record.nil? && current_user.id == @item.user_id
       if @item.destroy
         redirect_to root_path
       end
